Makefile: add library for mkfs.btrfs
Even if mkfs.btrfs can be executed from a shell, there are conditions
(see the reasons for the creation of libbtrfsutil) to call a function
inside own applicatiuon code (security, better control, etc.).

Create a libmkfsbtrfs library with min_mkfs as entry point and the same
syntax like mkfs.btrfs. This can be linked to applications requiring to
create the filesystem.

Signed-off-by: Stefano Babic <sbabic@denx.de>